update Widget for Transaction Page
