+++ lib/app/modules/transaction/views/transaction_view.dart
@@ -4,6 +4,7 @@ import 'package:get/get.dart';
 import 'package:temanbicara/app/modules/transaction_method/views/transaction_method_view.dart';
 import 'package:temanbicara/app/themes/colors.dart';
 import 'package:temanbicara/app/themes/fonts.dart';
+import 'package:temanbicara/app/widgets/transaction/transactionData.dart';
 
 import '../controllers/transaction_controller.dart';
